Force full screen presentation for Setup View Controllers
In iOS 13, modal view controllers are by default shown as a card stack. This is cool, but makes them auto-dismissible and doesn’t call the underlying controller’s `viewDidAppear` method again when the VC dismisses.

This clashes with out implementation for showing the setup VCs in order and causes #15.

This PR forces the VCs to always present as full screen, restoring the previous functionality.

Fixes #15